pyproject.toml
[project]
name = "retrnote"
version = "0.1.0"
description = "Minimalist retro-inspired HTML tag-based word processor. Distraction free writing.."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
dependencies = [
	"beautifulsoup4>=4.12",
	"python-docx>=1.1",
	"lxml>=5.0"
]

[project.scripts]
retrnote = "retrnote.cli:main"

[project.urls]
Homepage = "https://github.com/vizeireann/retrnote"
Repository = "https://github.com/vizeireann/retrnote"
